+++ lib/osm/badge.rb
@@ -7,7 +7,7 @@ module Osm
     #   @return [String] the name of the badge
     # @!attribute [rw] requirement_notes
     #   @return [String] a description of the badge
-    # @!attribute [rw] key
+    # @!attribute [rw] osm_key
     #   @return [String] the key for the badge in OSM
     # @!attribute [rw] sections_needed
     #   @return [Fixnum]
@@ -45,13 +45,15 @@ module Osm
     # Get badges
     # @param [Osm::Api] api The api to use to make the request
     # @param [Osm::Section, Fixnum, #to_i] section The section (or its ID) to get the due badges for
+    # @param [Symbol] section_type The type of section to get badges for (if nil uses the type of the section param)
     # @!macro options_get
     # @return [Array<Osm::Badge>]
-    def self.get_badges_for_section(api, section, options={})
-      raise Error, 'This method must be called on one of the subclasses (CoreBadge, ChallengeBadge, StagedBadge or ActivityBadge)' if badge_type.nil?
+    def self.get_badges_for_section(api, section, section_type=nil, options={})
+      raise Error, 'This method must be called on one of the subclasses (CoreBadge, ChallengeBadge, StagedBadge or ActivityBadge)' if type.nil?
       require_ability_to(api, :read, :badge, section, options)
       section = Osm::Section.get(api, section, options) unless section.is_a?(Osm::Section)
-      cache_key = ['badges', section.type, badge_type]
+      section_type ||= section.type
+      cache_key = ['badges', section_type, type]
 
       if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key)
         return cache_read(api, cache_key)
@@ -60,7 +62,7 @@ module Osm
       term_id = Osm::Term.get_current_term_for_section(api, section, options).to_i
       badges = []
 
-      data = api.perform_query("challenges.php?action=getInitialBadges&type=#{badge_type}&sectionid=#{section.id}&section=#{section.type}&termid=#{term_id}")
+      data = api.perform_query("challenges.php?action=getInitialBadges&type=#{type}&sectionid=#{section.id}&section=#{section_type}&termid=#{term_id}")
       badge_order = data["badgeOrder"].to_s.split(',')
       structures = data["structure"] || {}
       details = data["details"] || {}
@@ -97,34 +99,70 @@ module Osm
       return badges
     end
 
+    # Get a summary of badges earnt by members
+    # @param [Osm::Api] api The api to use to make the request
+    # @param [Osm::Section, Fixnum, #to_i] section The section (or its ID) to get the due badges for
+    # @param [Osm::Term, Fixnum, #to_i, nil] term The term (or its ID) to get the due badges for, passing nil causes the current term to be used
+    # @!macro options_get
+    # @return [Array<Hash>]
+    def self.get_summary_for_section(api, section, term=nil, options={})
+      raise Error, 'This method must be called on one of the subclasses (CoreBadge, ChallengeBadge, StagedBadge or ActivityBadge)' if type.nil?
+      require_ability_to(api, :read, :badge, section, options)
+      section = Osm::Section.get(api, section, options) unless section.is_a?(Osm::Section)
+      term_id = (term.nil? ? Osm::Term.get_current_term_for_section(api, section, options) : term).to_i
+      cache_key = ['badge-summary', section.id, term_id, type]
+
+      if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key)
+        return cache_read(api, cache_key)
+      end
+
+      summary = []
+      data = api.perform_query("challenges.php?action=summary&section=#{section.type}&sectionid=#{section.id}&termid=#{term_id}&type=#{type}")
+      data['items'].each do |item|
+        new_item = {
+          :first_name => item['firstname'],
+          :last_name => item['lastname'],
+        }
+        (item.keys - ['firstname', 'lastname']).each do |key|
+          new_item[key] = item[key]
+        end
+        summary.push new_item
+      end
+
+      cache_write(api, cache_key, summary)
+      return summary
+    end
+
     # Get a list of badge requirements met by members
     # @param [Osm::Api] api The api to use to make the request
     # @param [Osm::Section, Fixnum, #to_i] section The section (or its ID) to get the due badges for
-    # @param [Osm::Badge] badge The badge to get data for
     # @param [Osm::Term, Fixnum, #to_i, nil] term The term (or its ID) to get the due badges for, passing nil causes the current term to be used
     # @!macro options_get
     # @return [Array<Osm::Badge::Data>]
-    def self.get_badge_data_for_section(api, section, badge, term=nil, options={})
-      raise Error, 'This method must be called on one of the subclasses (CoreBadge, ChallengeBadge, StagedBadge or ActivityBadge)' if badge_type.nil?
+    def get_data_for_section(api, section, term=nil, options={})
+      raise Error, 'This method must be called on one of the subclasses (CoreBadge, ChallengeBadge, StagedBadge or ActivityBadge)' if type.nil?
       Osm::Model.require_ability_to(api, :read, :badge, section, options)
       section = Osm::Section.get(api, section, options) unless section.is_a?(Osm::Section)
       term_id = (term.nil? ? Osm::Term.get_current_term_for_section(api, section, options) : term).to_i
-      cache_key = ['badge_data', section.id, term_id, badge.osm_key]
+      cache_key = ['badge_data', section.id, term_id, osm_key]
 
       if !options[:no_cache] && cache_exist?(api, cache_key)
         return cache_read(api, cache_key)
       end
 
       datas = []
-      data = api.perform_query("challenges.php?termid=#{term_id}&type=#{badge_type}&section=#{section.type}&c=#{badge.osm_key}&sectionid=#{section.id}")
+      data = api.perform_query("challenges.php?termid=#{term_id}&type=#{type}&section=#{section.type}&c=#{osm_key}&sectionid=#{section.id}")
       data['items'].each do |d|
         datas.push Osm::Badge::Data.new(
           :member_id => d['scoutid'],
-          :completed => d['completed'].eql?('1'),
+          :first_name => d['firstname'],
+          :last_name => d['lastname'],
+          :completed => d['completed'].to_i,
+          :awarded => d['awarded'].to_i,
           :awarded_date => Osm.parse_date(d['awardeddate']),
           :requirements => d.select{ |k,v| k.include?('_') },
           :section_id => section.id,
-          :badge => badge,
+          :badge => self,
         )
       end
 
@@ -140,12 +178,19 @@ module Osm
     end
 
 
-    private
-    def self.badge_type
+    def self.type
       nil
     end
+    def type
+      self.class.type
+    end
+
+    private
     def self.subscription_required
       :bronze
+    end
+    def subscription_required
+      self.class.subscription_required
     end
 
 
@@ -196,14 +241,17 @@ module Osm
     end # Class Requirement
 
 
-    class Data
-      include ::ActiveAttr::MassAssignmentSecurity
-      include ::ActiveAttr::Model
-
+    class Data < Osm::Model
       # @!attribute [rw] member_id
       #   @return [Fixnum] ID of the member this data relates to
+      # @!attribute [rw] first_name
+      #   @return [Fixnum] the member's first name
+      # @!attribute [rw] last_name
+      #   @return [Fixnum] Ithe member's last name
       # @!attribute [rw] completed
-      #   @return [Boolean] whether this badge has been completed (i.e. it is due?)
+      #   @return [Fixnum] whether this badge has been completed (i.e. it is due?), number indicates stage if appropriate
+      # @!attribute [rw] awarded
+      #   @return [Date] the last stage awarded
       # @!attribute [rw] awarded_date
       #   @return [Date] when the badge was awarded
       # @!attribute [rw] requirements
@@ -214,16 +262,22 @@ module Osm
       #   @return [Osm::Badge] the badge that the data belongs to
 
       attribute :member_id, :type => Integer
-      attribute :completed, :type => Boolean
-      attribute :awarded_date, :type => Date
+      attribute :first_name, :type => String
+      attribute :last_name, :type => String
+      attribute :completed, :type => Integer, :default => 0
+      attribute :awarded, :type => Integer, :default => 0
+      attribute :awarded_date, :type => Date, :default => nil
       attribute :requirements, :type => Object, :default => DirtyHashy.new
       attribute :section_id, :type => Integer
       attribute :badge, :type => Object
 
-      attr_accessible :member_id, :completed, :awarded_date, :requirements, :section_id, :badge
+      attr_accessible :member_id, :first_name, :last_name, :completed, :awarded, :awarded_date, :requirements, :section_id, :badge
 
       validates_presence_of :badge
-      validates_inclusion_of :completed, :in => [true, false]
+      validates_presence_of :first_name
+      validates_presence_of :last_name
+      validates_numericality_of :completed, :only_integer=>true, :greater_than_or_equal_to=>0
+      validates_numericality_of :awarded, :only_integer=>true, :greater_than_or_equal_to=>0
       validates_numericality_of :member_id, :only_integer=>true, :greater_than=>0
       validates_numericality_of :section_id, :only_integer=>true, :greater_than=>0
       validates :requirements, :hash => {:key_type => String, :value_type => String}
@@ -246,7 +300,7 @@ module Osm
       def total_gained
         count = 0
         requirements.each do |field, data|
-          next if data.blank? || data.downcase[0].eql?('x')
+          next if data.blank? || data[0].downcase.eql?('x')
           count += 1
         end
         return count
@@ -273,10 +327,108 @@ module Osm
         requirements.each do |field, data|
           field = field.split('_')[0]
           count[field] ||= 0
-          next if data.blank? || data.downcase[0].eql?('x')
+          next if data.blank? || data[0].downcase.eql?('x')
           count[field] += 1
         end
         return count
+      end
+
+      # Check if this badge is due
+      # @return [Boolean] whether the badge is due to the member
+      def due?
+        completed > awarded
+      end
+
+      # Check if this badge has been started
+      # @return [Boolean] whether the badge has been started by the member (always false if the badge has been completed)
+      def started?
+        unless badge.type == :staged
+          return false if completed?
+          requirements.each do |key, value|
+            return true unless value.blank? || value[0].downcase.eql?('x')
+          end
+        else
+          # Staged badge
+          return (started > completed)
+        end
+        return false
+      end
+
+      # Get which stage has been started
+      # @return [Fixnum] which stage of the badge has been started by the member (lowest)
+      def started
+        unless badge.type == :staged
+          return started? ? 1 : 0
+        else
+          # Staged badge
+          if ['nightsaway', 'hikes'].include?(badge.osm_key) # Special staged badges
+            stages = [1, 5, 10, 20, 35, 50, 75, 100, 125, 150, 175, 200] if badge.osm_key.eql?('nightsaway')
+            stages = [1, 5, 10, 20, 35, 50] if badge.osm_key.eql?('hikes')
+            done = requirements['y_01'].to_i
+            return 0 if done < stages[0]                # Not started the first stage
+            return 0 if done >= stages[stages.size - 1] # No more stages can be started
+            (1..stages.size-1).reverse_each do |index|
+              if (done < stages[index]) && (done > stages[index-1])
+                return stages[index]
+              end
+            end
+          else
+            start_group = 'abcde'[completed] # Requirements use the group letter to denote stage
+            started = 'z'
+            requirements.each do |key, value|
+              next if key[0] < start_group # This stage is marked as completed
+              next if key[0] > started     # This stage is after the stage currently started
+              started = key[0] unless value.blank? || value[0].downcase.eql?('x')
+            end
+            return started.eql?('z') ? 0 : 'abcde'.index(started)+1
+          end
+          return 0
+        end
+      end
+
+      # Mark the badge as awarded in OSM
+      # @param [Osm::Api] api The api to use to make the request
+      # @param [Date] date The date to mark the badge as awarded
+      # @param [Fixnum] level The level of the badge to award (1 for non-staged badges)
+      # @param [Symbol] mark_as :awarded or :due
+      # @return [Boolean] whether the data was updated in OSM
+      def mark_awarded(api, date=Date.today, level=completed, mark_as=:awarded)
+        raise ArgumentError, 'date is not a Date' unless date.is_a?(Date)
+        raise ArgumentError, 'mark_as is not an allowed value, use :awarded or :du' unless [:awarded, :due].include?(mark_as)
+        raise ArgumentError, 'level can not be negative' if level < 0
+        section = Osm::Section.get(api, section_id)
+        require_ability_to(api, :write, :badge, section)
+
+        date_formatted = date.strftime(Osm::OSM_DATE_FORMAT)
+
+        result = api.perform_query("challenges.php?action=award", {
+          'dateAwarded' => date_formatted,
+          'sectionid' => section_id,
+          'section' => section.type,
+          'chal' => badge.osm_key,
+          'type' => badge.type,
+          'stagedLevel' => level,
+          'due' => mark_as,
+        })
+        updated = result.is_a?(Array) &&
+                  result[0].is_a?(Hash) &&
+                  (result[0]['sid'].to_i == member_id) &&
+                  (result[0]['awarded'].to_i == level) &&
+                  (result[0]['awardeddate'] == date_formatted)
+
+        if updated
+          awarded = level
+          awarded_date = date
+        end
+        return updated
+      end
+
+      # Mark the badge as due in OSM
+      # @param [Osm::Api] api The api to use to make the request
+      # @param [Fixnum] level The level of the badge to mark as due (1 for non-staged badges)
+      # @return [Boolean] whether the data was updated in OSM
+      def mark_due(api, level)
+        mark_awarded(api, Date.today, level, :due)
       end
 
       # Update data in OSM
@@ -286,13 +438,13 @@ module Osm
       def update(api)
         raise Osm::ObjectIsInvalid, 'data is invalid' unless valid?
         section = Osm::Section.get(api, section_id)
-        Osm::Model.require_ability_to(api, :write, :badge, section)
+        require_ability_to(api, :write, :badge, section)
 
         updated = true
         editable_fields = badge.requirements.select{ |r| r.editable }.map{ |r| r.field}
         requirements.changes.each do |field, (was,now)|
           if editable_fields.include?(field)
-            result = api.perform_query("challenges.php?type=#{badge.class.badge_type}&section=#{section.type}", {
+            result = api.perform_query("challenges.php?type=#{badge.class.type}&section=#{section.type}", {
               'action' => 'updatesingle',
               'id' => member_id,
               'col' => field,
@@ -306,9 +458,16 @@ module Osm
           end
         end
 
-
         if updated
           requirements.clean_up!
+        end
+
+        if changed_attributes.include?('awarded') || changed_attributes.include?('awarded_date')
+          if mark_awarded(api, awarded_date, awarded)
+            reset_changed_attributes
+          else
+            updated = false
+          end
         end
 
         return updated
@@ -333,28 +492,28 @@ module Osm
 
   class CoreBadge < Osm::Badge
     private
-    def self.badge_type
+    def self.type
       :core
     end
   end # Class CoreBadge
 
   class ChallengeBadge < Osm::Badge
     private
-    def self.badge_type
+    def self.type
       :challenge
     end
   end # Class ChallengeBadge
 
   class StagedBadge < Osm::Badge
     private
-    def self.badge_type
+    def self.type
       :staged
     end
   end # Class StagedBadge
 
   class ActivityBadge < Osm::Badge
     private
-    def self.badge_type
+    def self.type
       :activity
     end
     def self.subscription_required