+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## Version 0.6.0
+
+  * Badge::Data
+    * completed attribute is now a Fixnum not Boolean (fixes staged badges)
+    * Add awarded attribute (the last level awarded)
+    * Add due? method to tell if the badge is due
+    * Add started? method to tell if the badge has been started
+    * Add started method to tell which stage has been started
+    * Add first\_name and last\_name attributes
+    * Add mark\_awarded method
+    * Add mark\_due method
+  * Badge
+    * Make get\_badge\_data\_for\_section an instance not class method (called get\_data\_for\_section)
+    * Add get\_summary\_for\_section(api, section, term=nil, options={}) method
+    * Add optional section\_type parameter to get\_badges\_for\_section method
+
 ## Version 0.5.0
 
   * Code breaking changes to DueBadges: